requirements.txt
fasthtml
pillow
pytesseract
PyPDF2
python-pptx
openai
httpx
edge-tts